requirements-dev.txt
# CIRISBridge Development Dependencies

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0.0

# Linting
ansible-lint>=6.0.0

# Type checking (optional)
# mypy>=1.0.0
